plugins/discovery: Fix log drop flaky test (#3373)
The log drop check test does not setup a remote
decision logging endpoint to upload logs. Hence when the
decision log plugin tries to upload a log, it fails. Now when
the plugin tries to requeue the log it causes the log
drop count to increase as the rate limit has already exceeded.

This change updates the test to take into account the drop count
increase caused due to such a scenario.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>