wasm-sdk: relax validations, only provide needed imports (#3449)
* wasm+sdk: wasm-opt related changes

To enable using this for benchmarks, we can now skip the warning.

Also, when the wasm is passed through wasm-opt, some of its imports
may get dropped. This changes the wasm sdk code to allow for modules
that only have a subset of the expected imports.

* wasm-sdk: allow for using `opa_println()` in wasm modules

This is for debugging and development only. But having changed the
imports to allow for a subset, we can finally have it that way.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>