fix nightly fuzz build (#3474)
* workflow: add 'quick fuzz' run

This is just a very short (3m) fuzz check, to ensure that we don't
surprise ourselves when merging something that in some weird way
affects the nightly fuzzer run. It's probably unlikely to come up
with interesting results, but that's what the long nightly run is for.

* build/fuzzer: fix go module state

Something went amiss here. It seems like running these two commands
makes the build work again:

    $ go mod tidy
    $ go get github.com/dvyukov/go-fuzz/go-fuzz-dep
    go get: added github.com/dvyukov/go-fuzz v0.0.0-20210429054444-fca39067bc72

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>