Run dockerized golangci-lint
This means that "make lint" will work the same for everyone and the
workflow without additional coordination. Before this PR, it was
possible for maintainers to be on a different version of golangci-lint
than the one used by the GitHub workflow and so "make check" would
provide inconsistent results.

Also add timeout to configuration so we don't time out.`

Signed-off-by: Will Beason <willbeason@google.com>