topdown: Fix PE unknown check to avoid saving unnecessarily
The PE unknown check in the evalTree phase while finishing a ref was
incorrectly checking for unknowns on the unplugged ref. This mean that
if a ref element was unknown, that the expression would get saved and
evaluation would be skipped even though the ref key was
enumerated. This resulted in more results being generated by PE than
necessary and lead to expressions that could not be indexed.

Fixes #3552

Signed-off-by: Torin Sandall <torinsandall@gmail.com>