docs: Add guidance for OPA-Envoy benchmarks
Previously we provided actual performance benchmarks
for the OPA-Envoy plugin by running it in an experimental
setup. Although it maybe useful to document real numbers,
the results depend a lot on the environment in which OPA-Envoy
is run, OPA features being used and also on the methodology
used to generate them (eg. how a particular load test tool
measures latency).

Hence this commit provides guidance and best practices for
benchmarking the OPA-Envoy plugin which users can adopt in
their own environments.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>