docs/extensions: marshal the decision log (#3583)
I think this helps people who are not as familiar with go to get
more expected results. The issue is that because `input` and `result`
are interfaces when `event` is printed to stdout without the map being
marshalled to JSON, both `input` and `result` are printed as memory
pointers. This is confusing to someone less familiar with go as
they may believe that their input is not present in the decision
log. By adding Marshal, the stdout will now be standard JSON which
will then have `result` and `input` with the string values.

Signed-off-by: John Snow <john@styra.com>