workflow/pull_request: run tests on macos (#3795)
This splits off the generate call, so that we can have the wasm bits (and capabilities.json) from the PR used in the PR checks.

The Perf check still is the one taking longest, even with the added matrix build job and the jobs depending on it. The test matrix job was split off to run those in parallel, the binary smoke tests for example can already proceed.

To simplify things, we're relying on the setup-go action for both the linux and the darwin unit tests. (We could also use it for the builds of linux and windows binaries... but there, I'm more concerned about a clean build env and reproducibility.)

Fixes #3176.

Signed-off-by: Stephan Renatus <stephan.renatus@gmail.com>