compile: Add debug message if optimized entrypoint is discarded (#4030)
Previously, `opa build -O=1` would silently discard the support
module for the entrypoint if the entrypoint could not be PE-ed. This
could be somewhat confusing if users expect an optimized/ directory to
show up in the bundle. This change just adds a debug message to the
compiler so that we can see when this happens.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>